requirements.txt
requests
Pillow
numpy
scikit-image
geopy
torch
tensorflow
GDAL
rasterio
